norminee'd raycasting, added file to draw ray's screen column

+++ src/0_main.c
@@ -6,7 +6,7 @@
 /*   By: simon <simon@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/31 18:33:20 by svan-hoo          #+#    #+#             */
-/*   Updated: 2024/09/11 14:57:27 by simon            ###   ########.fr       */
+/*   Updated: 2024/09/12 16:49:40 by simon            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,8 +28,8 @@ static void
 	camera->movement_speed = MOVEMENT_SPEED * window->deltatime;
 	if (camera->movement_speed > 0.49)
 		camera->movement_speed = 0.49;
-	camera->rotation_cosin[0] = cos(ROTATION_SPEED * window->deltatime);
-	camera->rotation_cosin[1] = sin(ROTATION_SPEED * window->deltatime);
+	camera->rotation_cosin[0] = cosf(ROTATION_SPEED * window->deltatime);
+	camera->rotation_cosin[1] = sinf(ROTATION_SPEED * window->deltatime);
 }
 
 static void
